Feature development for self reporting absence.
Added checkbox to add availablebeforesession field in all the required pages.
If availablebeforesession is enabled in a status of a session then the global setting of studentscanmarksessiontime will be overridden.
Students can see and submit attendance for a status of a session which has availablebeforesession enabled
availablebeforesession allow attendance only on future sessions and not past.